v2.0.0: remove counter
Removed the need for an atomic counter for the commitId by combining
the committedAt and aggregateId properties. This assumes that a single
aggregate will not have more than one commit per millisecond. In practice
this should hold true as a version conflict would likely occur. In
addition, callbacks were removed and promises used instead. The AWS
object has been exposed to allow for custom Promise implementations.